move ProfiledComponent into util dir
